fix: harden mutable proxy async context refresh from review findings
Review fixes on top of #6691:

- Treat structural changes (deleted key, removed attr, changed container
  type) during proxy refresh as the documented RuntimeError instead of
  leaking raw KeyError/AttributeError/TypeError.
- Rename the "iter" path sentinel to "unrefreshable" (it also marks
  missing dict.get defaults) and reject unrefreshable proxies before
  taking the state lock; mark list index/slice proxies unrefreshable
  since positions are not stable identities.
- Catch BaseException in __aenter__ so KeyboardInterrupt cannot leak
  the state lock; drop the IterPath dynamic-subclass optimization.
- Skip the wrapping machinery for non-mutable element reads and split
  _wrap_recursive/_wrap_mutable so mutability is checked exactly once:
  dict element reads ~4.5x faster, list iteration ~8.6x faster than
  main; iterating all-mutable elements ~13% slower from path tracking.
- Move mutable proxy tests to tests/units/istate/test_proxy.py and add
  coverage: attr paths, plain-state no-op enter, deleted-key refresh,
  shifted list items, BaseException cleanup.
- Document refresh-by-path semantics and unrefreshable cases in the
  background events docs.
